Format string fixes to correctly compile on 64-bit systems.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/socketapi@1334 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ cppsocket_programs/sctpmultiserver.cc
@@ -362,7 +362,7 @@ void CharGenServer::run()
 
    for(;;) {
       // ====== Generate string =============================================
-      snprintf((char*)&dataBuffer,256,"#%08Ld: ",line++);
+      snprintf((char*)&dataBuffer,256,"#%08llu: ",(unsigned long long)line++);
       integer i,j;
       j = strlen(dataBuffer);
       for(i = 30;i < 128;i++) {
